// serviceLayer/src/fbs/generated/model/holdingsForBibliographicalRecordLogisticsV1.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * FBS Adapter
 * OpenAPI spec version: 1.0
 */
import type { HoldingsLogisticsV1 } from "./holdingsLogisticsV1";

export interface HoldingsForBibliographicalRecordLogisticsV1 {
  /** An array of holdings for the materials matching the bibliographical record, as distributed across branches,
 departments and locations */
  holdings: HoldingsLogisticsV1[];
  /** Identifies the bibliographical record for the available materials,
 The FAUST number */
  recordId: string;
  /** True if there is any reservable materials */
  reservable: boolean;
  /** Total number of current active reservations for the bibliographical record */
  reservations: number;
}